root|tooling: Remove violation-free rules from deferred rules
