+++ .config/strict-lint.config.ts
@@ -1,6 +1,6 @@
 import type { StrictLintConfig } from '@williamthorsen/strict-lint';
 
-import { deferredLintRules } from './deferred-lint-rules.ts';
+import { deferredLintRules } from './eslint/deferred-lint-rules.ts';
 
 const config: StrictLintConfig = {
   // Keep the deferred rules as warnings; strict-lint otherwise promotes every warning to an error.
