+++ eslint.config.ts
@@ -1,7 +1,7 @@
 import baseConfig from '@williamthorsen/eslint-config-typescript';
 import { defineConfig } from 'eslint/config';
 
-import { deferredLintRules } from './.config/deferred-lint-rules.ts';
+import { deferredLintRules } from './.config/eslint/deferred-lint-rules.ts';
 
 const config = defineConfig([
   ...baseConfig,
